requirements.txt
opencv-python
PyYAML